change to single filter

+++ project1/src/gpu/cuda_PartB.cu
@@ -3,7 +3,7 @@
 #include "utils.hpp"
 
 const int FILTER_SIZE = 3;
-__constant__ double filter[FILTER_SIZE][FILTER_SIZE] = {
+__constant__ float filter[FILTER_SIZE][FILTER_SIZE] = {
   { 1.0 / 9, 1.0 / 9, 1.0 / 9 },
   { 1.0 / 9, 1.0 / 9, 1.0 / 9 },
   { 1.0 / 9, 1.0 / 9, 1.0 / 9 }
